Refactor Collections::Edit::PaneComponent
Since this component just renders another template with an adjustment
to a single slot, this refactors it to inherit the other component
and take advantage of template inheritance.

This allows you to use Collections::Edit::PaneComponent more like
its parent, including passing the "help" slot, and removes the
need for a separate template.

Needed to support #274, because the Terms of Use pane has a very
long help text that needs to be rendered using the "help" slot.