Ensure custom terms of use entry instructions are populated
This implements a change discussed at standup that sets the
instructions for entering custom terms of use at import time,
so that they are never empty after migrating data from H2.

Previously, we were substituting empty instructions with a
default value at render time in the Work form. It was possible
to have selected 'depositor enters terms of use' but not
provide any instructions for doing that.

This also allows for renaming the field to something slightly
shorter and easier to remember.

Prep for #274, where we want to make this field in the
Collection form required and also pre-populate it with our
default.