Standardize on druid as param for collection and work paths
Fixes #958

It remains possible to link directly to a collection or work database ID, which we do in a handful of situations where we want to go directly to the "wait" page. Instead of special-casing the druid paths/URLs which are the norm in H3, switch the behavior to special-case the oddballs.