Adjust accessible label for invalid tabs via aria-labelledby
Fixes #908

This alternative approach to #1038 adds a hidden element with
text indicating that a tab is invalid, and then modifies the
tab's label using aria-labelledby so that its name is a combination
of the tab name and the invalid description.

Tested in VO and confirmed it reads something like "Details, Error:
required fields missing, tab, active, 1 of 10".