Decouple the Notifier from mailer/action implementation details
With this change, `Notifier.subscribe` takes an event name to subscribe to, and then yields event payloads to the block. This means the Notifier no longer needs to care about whether something is a mailer class or an "action," terminology we no longer use in H3. This change makes the Notifier class less coupled to implementation details and gives us more flexibility to new use cases going forward.