Remove left/right padding from icon buttons
Fixes #335

This commit effectively reduces the amount of horizontal padding between icon buttons and adjacent fields, having the desired effect of bringing the field delete button and the text field for repeatable components (e.g., contact email) closer. It also removes the horizontal padding of the move up and move down buttons, which I checked out in the work edit screen for authors, and it does not meaningfully impact that part of the UI as the delete and move buttons are arranged vertically there.